+++ app/components/Link.react.js
@@ -1,26 +1,25 @@
+import R from 'ramda';
 import React from 'react';
-import * as styles from './Configuration.css';
 import {dynamicRequireElectron} from '../utils/utils';
+import * as styles from './Link.css';
+
 
 let shell = null;
 try {
     shell = dynamicRequireElectron().shell;
 } catch (e) {}
 
-export function Link(href, text) {
-
+export function Link(props) {
+    const {href} = props;
+    if (shell) {
         return (
-            <span>
-                {shell ? (
-                    <span className={styles.externalLink}
-                        onClick={() => {
-                            shell.openExternal(href);
-                    }}
-                    >
-                        {text}
-                    </span>
-                ) : <a className={styles.externalLink} href={href} target="_blank">{text}</a>}
-            </span>
-
+            <span
+                className={`${styles.a} ${props.className}`}
+                onClick={() => {shell.openExternal(href);}}
+                {...R.omit(['className'], props)}
+            />
         );
+    } else {
+        return <a href={href} target="_blank" {...props}/>
+    }
 }