+++ app/store/configureStore.development.js
@@ -25,10 +25,11 @@ export default function configureStore(initialState) {
     const store = createStore(rootReducer, initialState, enhancer);
 
     if (module.hot) {
-        module.hot.accept('../reducers', () =>
-        store.replaceReducer(require('../reducers'))
-    );
-}
+            module.hot.accept('../reducers', () =>
+            store.replaceReducer(require('../reducers'))
+        );
+    }
 
-return store;
+    window.store = store;
+    return store;
 }